second commit

+++ app/Http/Controllers/ReadJsonController.php
@@ -18,22 +18,37 @@ class ReadJsonController extends Controller
             /**
              * file_get_contents() is using for getting context from any file.
              */
-
-            $json = file_get_contents(base_path('public/simple.json'));
-            $json = json_decode($json, true);
-            $json['sandy'] = [
-               "name"=>"sandeep gour",
-               "email"=>"sandeep@gmail.com",
-               "mobile"=>"9087654321",
-               "dob"=>"02-11-1998"
-            ];
-
-            /**
-             * file_put_contents using for replace context in exist file or append context in file
-             */
-            $file = file_put_contents(base_path('public/simple.json'), json_encode($json), FILE_APPEND | LOCK_EX);
-            echo "<pre>";
-            print_r($json);
+            $json = [];
+            if (File::exists(public_path('simple.json'))) {
+                $json = file_get_contents(public_path('simple.json'));
+                $json = json_decode($json, true);
+                if ($json != null) {
+                    $data = [
+                        "name" => "sandeep gour",
+                        "email" => "sandeep@gmail.com",
+                        "mobile" => "9087654321",
+                        "dob" => "02-11-1998"
+                    ];
+                    array_push($json, $data);
+                    $newJsonString = json_encode($json);
+                    file_put_contents(public_path('simple.json'), $newJsonString);
+                } else {
+                    $data[0] = [
+                        "name" => "sandeep gour",
+                        "email" => "sandeep@gmail.com",
+                        "mobile" => "9087654321",
+                        "dob" => "02-11-1998"
+                    ];
+                    $newJsonString = json_encode($data);
+                    file_put_contents(public_path('simple.json'), $newJsonString);
+                }
+                $json_data = file_get_contents(public_path('simple.json'));
+                $json_data = json_decode($json_data, true);
+                echo "<pre>";
+                print_r($json_data);
+            } else {
+                echo "file is not exist!";
+            }
         } catch (\Exception $e) {
             return "file does't exist!";
         }
